requirements.txt
Django==5.0.2
gunicorn==20.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.2.0
django-filter
django-storages==1.13.2
google-cloud-storage
pillow
python-dotenv
nltk
pandas==2.2.2
numpy==1.26.4
boto3==1.26.59
gunicorn==20.1.0
psycopg2-binary==2.9.6
drf-yasg
dj_database_url
flake8
librosa
pydot
pgvector
torch
transformers
pillow